Update Name For Macedonia (#11)
Earlier this year the Macedonian Government renamed the country from
"Republic of Macedonia" to "Republic of North Macedonia". This change
doesn't automatically propagate into FreeAgent as we maintain our own
list of country names (a fork of the country-select gem). This PR
updates the country name on request of a customer (who has raised
several support requests and for which Support Engineering have been
manually updating invoices until now). (thanks to @dvjones89)